Replace FilterInterceptor with ControllerInterceptor for Rmodules plugin
